test(color-picker-web): refactor snapshots and tests for improved structure
