+++ packages/pluggableWidgets/color-picker-web/src/components/__tests__/ColorPicker.spec.tsx
@@ -1,5 +1,5 @@
 import "@testing-library/jest-dom";
-import { act, render, RenderResult } from "@testing-library/react";
+import { render, RenderResult } from "@testing-library/react";
 import userEvent, { UserEvent } from "@testing-library/user-event";
 import { createElement } from "react";
 import { ColorPicker, ColorPickerProps } from "../ColorPicker";
@@ -27,33 +27,6 @@ describe("ColorPicker", () => {
     function renderColorPicker(configs: Partial<ColorPickerProps> = {}): RenderResult {
         return render(<ColorPicker {...colorPickerProps} {...configs} />);
     }
-
-    it("renders the structure correctly", () => {
-        const colorPickerComponent = renderColorPicker();
-
-        expect(colorPickerComponent.asFragment()).toMatchSnapshot();
-    });
-
-    it("that is disabled renders with the structure", () => {
-        const colorPickerComponent = renderColorPicker({ disabled: true });
-        expect(colorPickerComponent.asFragment()).toMatchSnapshot();
-    });
-
-    it("renders picker with pre-defined default colors", () => {
-        const colorPickerComponent = renderColorPicker();
-
-        act(() => {
-            colorPickerComponent.rerender(
-                <ColorPicker
-                    {...colorPickerProps}
-                    mode="inline"
-                    defaultColors={[{ color: "#2CCCE4" }, { color: "#555555" }]}
-                />
-            );
-        });
-
-        expect(colorPickerComponent.asFragment()).toMatchSnapshot();
-    });
 
     it("should handle on change event", async () => {
         const { getByTitle } = renderColorPicker({
